Fix links for Native platforms
