+++ src/components/AnchorForCommentsOnly/index.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import * as anchorForCommentsOnlyPropTypes from './anchorForCommentsOnlyPropTypes';
+import BaseAnchorForCommentsOnly from './BaseAnchorForCommentsOnly';
+
+// eslint-disable-next-line react/jsx-props-no-spreading
+const AnchorForCommentsOnly = props => <BaseAnchorForCommentsOnly {...props} />;
+AnchorForCommentsOnly.propTypes = anchorForCommentsOnlyPropTypes.propTypes;
+AnchorForCommentsOnly.defaultProps = anchorForCommentsOnlyPropTypes.defaultProps;
+AnchorForCommentsOnly.displayName = 'AnchorForCommentsOnly';
+
+export default AnchorForCommentsOnly;
